fix(metamask): open unlock dialog
