+++ packages/plugins/EVM/src/state/Connection/interceptors/MetaMask.ts
@@ -1,8 +1,11 @@
-import { EthereumMethodType } from '@masknet/web3-shared-evm'
+import { EthereumMethodType, isRiskMethod } from '@masknet/web3-shared-evm'
 import type { Context, Middleware } from '../types'
 
 export class MetaMask implements Middleware<Context> {
     async fn(context: Context, next: () => Promise<void>) {
+        if (isRiskMethod(context.request.method as EthereumMethodType)) {
+            await context.connection.connect()
+        }
         switch (context.request.method) {
             case EthereumMethodType.PERSONAL_SIGN:
                 context.requestArguments = {
